StockAnalyzer does what it says.
This is supposed to analyze stocks. Right now it just passes the stock symbol to stock. Eventually it may import a list of stocks you want analyzed, and export analysis of each. Might just let the user put one in at a time to be a good robot. Probably the second one though.