Grabs data to fill stock
Stock is useless without information. This feeds the data into stock. Or at least it will. For now it just opens the page. Ideally we could just take all the info, but the webpage source code is just too much for me to parse right now so it will only be taking in the currently relevant information discluding the PE since that's not stored on the page segment I pull and will need to add it in later.